Rename file, remove 'i' prefix
