+++ src/injected/visualization/formatter.ts
