+++ src/injected/visualization/highlight-box-formatter.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { IAssessmentVisualizationInstance } from '../frameCommunicators/html-element-axe-results-helper';
 import { FailureInstanceFormatter } from './failure-instance-formatter';
-import { DrawerConfiguration } from './iformatter';
+import { DrawerConfiguration } from './formatter';
 import { IssuesFormatter } from './issues-formatter';
 
 export class HighlightBoxFormatter extends FailureInstanceFormatter {
