+++ src/injected/visualization/single-target-formatter.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { IFormatter, SingleTargetDrawerConfiguration } from './iformatter';
+import { IFormatter, SingleTargetDrawerConfiguration } from './formatter';
 
 export class SingleTargetFormatter implements IFormatter {
     constructor(private injectedClassName: string) {}
