+++ src/injected/visualization/svg-drawer-v2.ts
@@ -8,9 +8,9 @@ import { WindowUtils } from './../../common/window-utils';
 import { BaseDrawer } from './base-drawer';
 import { CenterPositionCalculator } from './center-position-calculator';
 import { DrawerUtils } from './drawer-utils';
+import { ISVGDrawerConfiguration } from './formatter';
 import { IDrawerInitData } from './idrawer';
 import { IFocusIndicator } from './ifocus-indicator';
-import { ISVGDrawerConfiguration } from './iformatter';
 import { IPoint } from './ipoint';
 import { ITabbedItem } from './itabbed-item';
 import { SVGNamespaceUrl } from './svg-constants';
