+++ src/injected/visualization/svg-shape-factory.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { DrawerUtils } from './drawer-utils';
-import { ICircleConfiguration, ILineConfiguration, IStrokeConfiguration, ITextConfiguration } from './iformatter';
+import { ICircleConfiguration, ILineConfiguration, IStrokeConfiguration, ITextConfiguration } from './formatter';
 import { IPoint } from './ipoint';
 import { SVGNamespaceUrl } from './svg-constants';
